meow 57

+++ esoragoto/views.py
@@ -460,8 +460,8 @@ def graph_n1(request):
 
 def graph_n2(request):
     N = int(request.POST['N'])
-    m = float(request.POST['m'])
-    t__ = float(request.POST['t'])
+    m = int(request.POST['m'])
+    t__ = int(request.POST['t'])
     K = linspace(0,t__,m)
     interval_t = 100  #пространство i
